Some tweaks to make the game "sort of" run on HTML5. Would need lots of work though.

+++ scripts/sv6_load.gml
@@ -4,15 +4,17 @@ load_character_vars();
 reset_map();
 init_map();
 filename = argument0;
-file_copy(filename, filename + "d");
-crypt(filename + "d", "XOR_DFJykQ8xX3PuNnkLt6QviqALOLF8cxIDx1D63DAdph4KGQ4rOJ7", 2);
-fid = file_text_open_read(filename + "d");
+if (os_is_native) {
+    file_copy(filename, filename + "d");
+    crypt(filename + "d", "XOR_DFJykQ8xX3PuNnkLt6QviqALOLF8cxIDx1D63DAdph4KGQ4rOJ7", 2);
+    fid = file_text_open_read(filename + "d");
+} else fid = file_text_open_read(filename);
 header = rc4(file_text_read_string(fid), "HEADER_KEY");
 file_text_readln(fid);
 if (header != "[AM2R SaveData V7.0]") {
     file_text_close(fid);
     show_message_async("Save Data Corrupted");
-    file_delete(filename + "d");
+    if (os_is_native) file_delete(filename + "d");
     file_delete(filename);
     room_goto(titleroom);
 } else {
@@ -33,7 +35,7 @@ if (header != "[AM2R SaveData V7.0]") {
     sv6_get_map(fid);
     file_text_readln(fid);
     file_text_close(fid);
-    file_delete(filename + "d");
+    if (os_is_native) file_delete(filename + "d");
     global.maxhealth = 99 + global.etanks * 100;
     global.playerhealth = global.maxhealth;
     if (global.difficulty < 2) {
